requirements/base.txt
# Django MSSQL
django-mssql
django-pymssql
pywin32
psycopg2
